renamed redux to store and deleted old files
