+++ client/web/src/components/MapPage/index.js
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
-import * as select from 'redux/selectors'
 import { useLocation } from 'react-router-dom'
-import { saveQuery, getJurisdiction } from 'redux/actions'
+import * as select from 'store/selectors'
+import { saveQuery, getJurisdiction } from 'store/actions'
 import useBreakpoints from 'hooks/useBreakpoints'
 import Desktop from './layouts/Desktop'
 import Mobile from './layouts/Mobile'
