--- client/web/src/redux-OLD/actions/data.js
@@ -1,12 +1,0 @@
-import api from 'services/api'
-
-export function getJurisdiction(jurisdictionId) {
-  return (dispatch) => {
-    return api.getJurisdiction(jurisdictionId).then((data) =>
-      dispatch({
-        type: 'GET_JURISDICTION_SUCCESS',
-        data,
-      })
-    )
-  }
-}
