--- client/web/src/redux-OLD/actions/search.js
@@ -1,3 +1,0 @@
-export function addSearch(search) {
-  return (dispatch) => dispatch({ type: 'ADD_SEARCH', search: search })
-}
