--- client/web/src/redux-OLD/actions/ui.js
@@ -1,4 +1,0 @@
-export function addSelectedLocation(location) {
-  return (dispatch) =>
-    dispatch({ type: 'ADD_SELECTED_LOCATION', location: location })
-}
