--- client/web/src/redux-OLD/preloadedState.js
@@ -1,12 +1,0 @@
-import chicagoParks from './chicago-parks.json'
-
-const preloadedState = {
-  data: null,
-  chicagoParks: chicagoParks,
-  searches: [],
-  ui: {
-    selectedLocation: null,
-  },
-}
-
-export default preloadedState
