--- client/web/src/redux-OLD/reducers/data.js
@@ -1,8 +1,0 @@
-export default (state = [], action) => {
-  switch (action.type) {
-    case 'GET_JURISDICTION_SUCCESS':
-      return action.data
-    default:
-      return state
-  }
-}
