--- client/web/src/redux-OLD/reducers/search.js
@@ -1,8 +1,0 @@
-export default (state = [], action) => {
-  switch (action.type) {
-    case 'ADD_SEARCH':
-      return [...state, action.search]
-    default:
-      return state
-  }
-}
