--- client/web/src/redux-OLD/reducers/ui.js
@@ -1,11 +1,0 @@
-export default (state = [], action) => {
-  switch (action.type) {
-    case 'ADD_SELECTED_LOCATION':
-      return {
-        ...state,
-        selectedLocation: action.location,
-      }
-    default:
-      return state
-  }
-}
