--- client/web/src/redux-OLD/rootReducer.js
@@ -1,12 +1,0 @@
-import { combineReducers } from 'redux'
-import searches from './reducers/search'
-import data from './reducers/data'
-import ui from './reducers/ui'
-import chicagoParks from './reducers/chicagoParks'
-
-export default combineReducers({
-  searches,
-  data,
-  ui,
-  chicagoParks,
-})
