--- client/web/src/redux-OLD/store.js
@@ -1,14 +1,0 @@
-import { createStore, applyMiddleware } from 'redux'
-import thunk from 'redux-thunk'
-import { composeWithDevTools } from 'redux-devtools-extension'
-
-import preloadedState from './preloadedState'
-import rootReducer from './rootReducer'
-
-const store = createStore(
-  rootReducer,
-  preloadedState,
-  composeWithDevTools(applyMiddleware(thunk))
-)
-
-export default store
